Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

SolverTools.AbstractTrustRegion

SolverTools.LineModel

SolverTools.TRONTrustRegion

SolverTools.TrustRegion

SolverTools.TrustRegionException

LinearOperators.reset!

NLPModels.grad

NLPModels.grad!

NLPModels.hess

NLPModels.obj

NLPModels.objgrad

NLPModels.objgrad!

SolverTools.acceptable

SolverTools.active

SolverTools.aredpred!

SolverTools.armijo_wolfe

SolverTools.breakpoints

SolverTools.compute_Hs_slope_qs!

SolverTools.hess!

SolverTools.project!

SolverTools.project_step!

SolverTools.redirect!

SolverTools.update!

​

SolverTools.AbstractTrustRegion — Type

AbstractTrustRegion

An abstract trust region type so that specific trust regions define update rules differently. Child types must have at least the following real fields:

acceptance_threshold

initial_radius

radius

ratio

and the following function:

update!(tr, step_norm)

SolverTools.LineModel — Type

A type to represent the restriction of a function to a direction. Given f : R → Rⁿ, x ∈ Rⁿ and a nonzero direction d ∈ Rⁿ,

ϕ = LineModel(nlp, x, d)

represents the function ϕ : R → R defined by

ϕ(t) := f(x + td).

SolverTools.TRONTrustRegion — Type

Trust region used by TRON

SolverTools.TrustRegion — Type

Basic trust region type.

SolverTools.TrustRegionException — Type

Exception type raised in case of error.

LinearOperators.reset! — Method

reset!(tr)

Reset the trust-region radius to its initial value

NLPModels.grad! — Method

grad!(f, t, g) evaluates the first derivative of the LineModel

ϕ(t) := f(x + td),

i.e.,

ϕ'(t) = ∇f(x + td)ᵀd.

The gradient ∇f(x + td) is stored in g.

NLPModels.grad — Method

grad(f, t) evaluates the first derivative of the LineModel

ϕ(t) := f(x + td),

i.e.,

ϕ'(t) = ∇f(x + td)ᵀd.

NLPModels.hess — Method

Evaluate the second derivative of the LineModel

ϕ(t) := f(x + td),

i.e.,

ϕ"(t) = dᵀ∇²f(x + td)d.

NLPModels.obj — Method

obj(f, t) evaluates the objective of the LineModel

ϕ(t) := f(x + td).

NLPModels.objgrad! — Method

objgrad!(f, t, g) evaluates the objective and first derivative of the LineModel

ϕ(t) := f(x + td),

and

ϕ'(t) = ∇f(x + td)ᵀd.

The gradient ∇f(x + td) is stored in g.

NLPModels.objgrad — Method

objgrad(f, t) evaluates the objective and first derivative of the LineModel

ϕ(t) := f(x + td),

and

ϕ'(t) = ∇f(x + td)ᵀd.

SolverTools.acceptable — Method

acceptable(tr)

Return true if a step is acceptable

SolverTools.active — Method

active(x, ℓ, u; rtol = 1e-8, atol = 1e-8)

Computes the active bounds at x, using tolerance min(rtol * (uᵢ-ℓᵢ), atol). If ℓᵢ or uᵢ is not finite, only atol is used.

SolverTools.aredpred! — Function

ared, pred = aredpred(tr, nlp, f, f_trial, Δm, x_trial, step, slope)

Compute the actual and predicted reductions ∆f and Δm, where Δf = f_trial - f is the actual reduction is an objective/merit/penalty function, Δm = m_trial - m is the reduction predicted by the model m of f. We assume that m is being minimized, and therefore that Δm < 0.

SolverTools.armijo_wolfe — Method

t, good_grad, ht, nbk, nbW = armijo_wolfe(h, h₀, slope, g)

Performs a line search from x along the direction d as defined by the LineModel $h(t) = f(x + t d)$, where h₀ = h(0) = f(x), slope = h'(0) = ∇f(x)ᵀd and g is a vector that will be overwritten with the gradient at various points. On exit, if good_grad=true, g contains the gradient at the final step length. The steplength is chosen trying to satisfy the Armijo and Wolfe conditions. The Armijo condition is

\[h(t) ≤ h₀ + τ₀ t h'(0)\]

and the Wolfe condition is

\[h'(t) ≤ τ₁ h'(0).\]

Initially the step is increased trying to satisfy the Wolfe condition. Afterwards, only backtracking is performed in order to try to satisfy the Armijo condition. The final steplength may only satisfy Armijo's condition.

The output is the following:

t: the step length;

good_grad: whether g is the gradient at x + t * d;

ht: the model value at t, i.e., f(x + t * d);

nbk: the number of times the steplength was decreased to satisfy the Armijo condition, i.e., number of backtracks;

nbW: the number of times the steplength was increased to satisfy the Wolfe condition.

The following keyword arguments can be provided:

t: starting steplength (default 1);

τ₀: slope factor in the Armijo condition (default max(1e-4, √ϵₘ));

τ₁: slope factor in the Wolfe condition. It should satisfy τ₁ > τ₀ (default 0.9999);

bk_max: maximum number of backtracks (default 10);

bW_max: maximum number of increases (default 5);

verbose: whether to print information (default false).

SolverTools.breakpoints — Method

nbrk, brkmin, brkmax = breakpoints(x, d, ℓ, u)

Find the smallest and largest values of α such that x + αd lies on the boundary. x is assumed to be feasible. nbrk is the number of breakpoints from x in the direction d.

SolverTools.compute_Hs_slope_qs! — Method

slope, qs = compute_Hs_slope_qs!(Hs, H, s, g)

Computes

Hs = H * s
slope = dot(g,s)
qs = ¹/₂sᵀHs + slope

SolverTools.hess! — Method

Evaluate the second derivative of the LineModel

ϕ(t) := f(x + td),

i.e.,

ϕ"(t) = dᵀ∇²f(x + td)d.

SolverTools.project! — Method

project!(y, x, ℓ, u)

Projects x into bounds ℓ and u, in the sense of yᵢ = max(ℓᵢ, min(xᵢ, uᵢ)).

SolverTools.project_step! — Method

project_step!(y, x, d, ℓ, u, α = 1.0)

Computes the projected direction y = P(x + α * d) - x.

SolverTools.redirect! — Method

redirect!(ϕ, x, d)

Change the values of x and d of the LineModel ϕ, but retains the counters.

SolverTools.update! — Function

update!(tr, step_norm)

Update the trust-region radius based on the ratio of actual vs. predicted reduction and the step norm.