API

API

API

Auxiliary

Benchmarking

Line-Search

Stats

Trust-Region

Auxiliary

SolverTools.active — Function.

active(x, ℓ, u; rtol = 1e-8, atol = 1e-8)

Computes the active bounds at x, using tolerance min(rtol * (uᵢ-ℓᵢ), atol). If ℓᵢ or uᵢ is not finite, only atol is used.

SolverTools.breakpoints — Function.

nbrk, brkmin, brkmax = breakpoints(x, d, ℓ, u)

Find the smallest and largest values of α such that x + αd lies on the boundary. x is assumed to be feasible. nbrk is the number of breakpoints from x in the direction d.

SolverTools.compute_Hs_slope_qs! — Function.

slope, qs = compute_Hs_slope_qs!(Hs, H, s, g)

Computes

Hs = H * s
slope = dot(g,s)
qs = ¹/₂sᵀHs + slope

SolverTools.log_header — Function.

log_header(colnames, coltypes)

Creates a header using the names in colnames formatted according to the types in coltypes. Uses internal formatting specification given by SolverTools.formats and default header translation given by SolverTools.default_headers.

Input:

colnames::Vector{Symbol}: Column names.

coltypes::Vector{DataType}: Column types.

Keyword arguments:

hdr_override::Dict{Symbol,String}: Overrides the default headers.

colsep::Int: Number of spaces between columns (Default: 2)

See also log_row.

SolverTools.log_row — Function.

log_row(vals)

Creates a table row from the values on vals according to their types. Pass the names and types of vals to log_header for a logging table. Uses internal formatting specification given by SolverTools.formats.

Keyword arguments:

colsep::Int: Number of spaces between columns (Default: 2)

SolverTools.project! — Function.

project!(y, x, ℓ, u)

Projects x into bounds ℓ and u, in the sense of yᵢ = max(ℓᵢ, min(xᵢ, uᵢ)).

SolverTools.project_step! — Function.

project_step!(y, x, d, ℓ, u, α = 1.0)

Computes the projected direction y = P(x + α * d) - x.

Benchmarking

SolverTools.bmark_solvers — Function.

bmark_solvers(solvers :: Dict{Symbol,Any}, args...; kwargs...)

Run a set of solvers on a set of problems.

Arguments

solvers: a dictionary of solvers to which each problem should be passed

other positional arguments accepted by solve_problems(), except for a solver name

Keyword arguments

Any keyword argument accepted by solve_problems()

Return value

A Dict{Symbol, AbstractExecutionStats} of statistics.

SolverTools.solve_problems — Function.

solve_problems(solver, problems :: Any; kwargs...)

Apply a solver to a set of problems.

Arguments

solver: the function name of a solver

problems: the set of problems to pass to the solver, as an iterable of AbstractNLPModel. It is recommended to use a generator expression (necessary for CUTEst problems).

Keyword arguments

solver_logger::AbstractLogger: logger wrapping the solver call. (default: NullLogger).

skipif::Function: function to be applied to a problem and return whether to skip it (default: x->false)

prune: do not include skipped problems in the final statistics (default: true)

any other keyword argument to be passed to the solver

Return value

a DataFrame where each row is a problem, minus the skipped ones if prune is true.

Line-Search

SolverTools.LineModel — Type.

A type to represent the restriction of a function to a direction. Given f : R → Rⁿ, x ∈ Rⁿ and a nonzero direction d ∈ Rⁿ,

ϕ = LineModel(nlp, x, d)

represents the function ϕ : R → R defined by

ϕ(t) := f(x + td).

NLPModels.obj — Function.

obj(f, t) evaluates the objective of the LineModel

ϕ(t) := f(x + td).

NLPModels.grad — Function.

grad(f, t) evaluates the first derivative of the LineModel

ϕ(t) := f(x + td),

i.e.,

ϕ'(t) = ∇f(x + td)ᵀd.

NLPModels.grad! — Function.

grad!(f, t, g) evaluates the first derivative of the LineModel

ϕ(t) := f(x + td),

i.e.,

ϕ'(t) = ∇f(x + td)ᵀd.

The gradient ∇f(x + td) is stored in g.

NLPModels.hess — Function.

Evaluate the second derivative of the LineModel

ϕ(t) := f(x + td),

i.e.,

ϕ"(t) = dᵀ∇²f(x + td)d.

SolverTools.redirect! — Function.

redirect!(ϕ, x, d)

Change the values of x and d of the LineModel ϕ, but retains the counters.

Stats

SolverTools.GenericExecutionStats — Type.

GenericExecutionStats(status, nlp; ...)

A GenericExecutionStats is a struct for storing output information of solvers. It contains the following fields:

status: Indicates the output of the solver. Use show_statuses() for the full list;

solution: The final approximation returned by the solver (default: []);

objective: The objective value at solution (default: Inf);

dual_feas: The dual feasibility norm at solution (default: Inf);

primal_feas: The primal feasibility norm at solution (default: 0.0 if uncontrained, Inf otherwise);

iter: The number of iterations computed by the solver (default: -1);

elapsed_time: The elapsed time computed by the solver (default: Inf);

counters::NLPModels.NLSCounters: The Internal structure storing the number of functions evaluations;

solver_specific::Dict{Symbol,Any}: A solver specific dictionary.

The counters variable is a copy of nlp's counters, and status is mandatory on construction. All other variables can be input as keyword arguments.

Notice that GenericExecutionStats does not compute anything, it simply stores.

SolverTools.show_statuses — Function.

show_statuses()

Show the list of available statuses to use with GenericExecutionStats.

Trust-Region

SolverTools.TrustRegionException — Type.

Exception type raised in case of error.

SolverTools.AbstractTrustRegion — Type.

AbstractTrustRegion

An abstract trust region type so that specific trust regions define update rules differently. Child types must have at least the following fields:

acceptance_threshold :: AbstractFloat

initial_radius :: AbstractFloat

radius :: AbstractFloat

ratio :: AbstractFloat

and the following function:

update!(tr, step_norm)

SolverTools.aredpred — Function.

ared, pred = aredpred(nlp, f, f_trial, Δm, x_trial, step, slope)

Compute the actual and predicted reductions ∆f and Δm, where Δf = f_trial - f is the actual reduction is an objective/merit/penalty function, Δm = m_trial - m is the reduction predicted by the model m of f. We assume that m is being minimized, and therefore that Δm < 0.

SolverTools.acceptable — Function.

acceptable(tr)

Return true if a step is acceptable

LinearOperators.reset! — Function.

reset!(tr)

Reset the trust-region radius to its initial value

SolverTools.get_property — Function.

A basic getter for AbstractTrustRegion instances. Should be overhauled when it's possible to overload getfield() and setfield!(). See https://github.com/JuliaLang/julia/issues/1974

SolverTools.set_property! — Function.

A basic setter for AbstractTrustRegion instances.

SolverTools.update! — Function.

update!(tr, step_norm)

Update the trust-region radius based on the ratio of actual vs. predicted reduction and the step norm.

SolverTools.TrustRegion — Type.

Basic trust region type.

SolverTools.TRONTrustRegion — Type.

Trust region used by TRON