API

API

Auxiliary

Line-Search

Stats

Trust-Region

Auxiliary

SolverTools.active — Function

active(x, ℓ, u; rtol = 1e-8, atol = 1e-8)

Computes the active bounds at x, using tolerance min(rtol * (uᵢ-ℓᵢ), atol). If ℓᵢ or uᵢ is not finite, only atol is used.

SolverTools.breakpoints — Function

nbrk, brkmin, brkmax = breakpoints(x, d, ℓ, u)

Find the smallest and largest values of α such that x + αd lies on the boundary. x is assumed to be feasible. nbrk is the number of breakpoints from x in the direction d.

SolverTools.compute_Hs_slope_qs! — Function

slope, qs = compute_Hs_slope_qs!(Hs, H, s, g)

Computes

Hs = H * s
slope = dot(g,s)
qs = ¹/₂sᵀHs + slope

SolverTools.log_header — Function

log_header(colnames, coltypes)

Creates a header using the names in colnames formatted according to the types in coltypes. Uses internal formatting specification given by SolverTools.formats and default header translation given by SolverTools.default_headers.

Input:

colnames::Vector{Symbol}: Column names.

coltypes::Vector{DataType}: Column types.

Keyword arguments:

hdr_override::Dict{Symbol,String}: Overrides the default headers.

colsep::Int: Number of spaces between columns (Default: 2)

See also log_row.

SolverTools.log_row — Function

log_row(vals)

Creates a table row from the values on vals according to their types. Pass the names and types of vals to log_header for a logging table. Uses internal formatting specification given by SolverTools.formats.

To handle a missing value, add the type instead of the number:

@info log_row(Any[1.0, 1])
@info log_row(Any[Float64, Int])

Prints

[ Info:  1.0e+00       1
[ Info:        -       -

Keyword arguments:

colsep::Int: Number of spaces between columns (Default: 2)

SolverTools.project! — Function

project!(y, x, ℓ, u)

Projects x into bounds ℓ and u, in the sense of yᵢ = max(ℓᵢ, min(xᵢ, uᵢ)).

SolverTools.project_step! — Function

project_step!(y, x, d, ℓ, u, α = 1.0)

Computes the projected direction y = P(x + α * d) - x.

Line-Search

SolverTools.LineModel — Type

A type to represent the restriction of a function to a direction. Given f : R → Rⁿ, x ∈ Rⁿ and a nonzero direction d ∈ Rⁿ,

ϕ = LineModel(nlp, x, d)

represents the function ϕ : R → R defined by

ϕ(t) := f(x + td).

NLPModels.obj — Function

obj(f, t) evaluates the objective of the LineModel

ϕ(t) := f(x + td).

NLPModels.grad — Function

grad(f, t) evaluates the first derivative of the LineModel

ϕ(t) := f(x + td),

i.e.,

ϕ'(t) = ∇f(x + td)ᵀd.

NLPModels.grad! — Function

grad!(f, t, g) evaluates the first derivative of the LineModel

ϕ(t) := f(x + td),

i.e.,

ϕ'(t) = ∇f(x + td)ᵀd.

The gradient ∇f(x + td) is stored in g.

NLPModels.objgrad — Function

objgrad(f, t) evaluates the objective and first derivative of the LineModel

ϕ(t) := f(x + td),

and

ϕ'(t) = ∇f(x + td)ᵀd.

NLPModels.objgrad! — Function

objgrad!(f, t, g) evaluates the objective and first derivative of the LineModel

ϕ(t) := f(x + td),

and

ϕ'(t) = ∇f(x + td)ᵀd.

The gradient ∇f(x + td) is stored in g.

NLPModels.hess — Function

Evaluate the second derivative of the LineModel

ϕ(t) := f(x + td),

i.e.,

ϕ"(t) = dᵀ∇²f(x + td)d.

SolverTools.redirect! — Function

redirect!(ϕ, x, d)

Change the values of x and d of the LineModel ϕ, but retains the counters.

SolverTools.armijo_wolfe — Function

t, good_grad, ht, nbk, nbW = armijo_wolfe(h, h₀, slope, g)

Performs a line search from x along the direction d as defined by the LineModel $h(t) = f(x + t d)$, where h₀ = h(0) = f(x), slope = h'(0) = ∇f(x)ᵀd and g is a vector that will be overwritten with the gradient at various points. On exit, if good_grad=true, g contains the gradient at the final step length. The steplength is chosen trying to satisfy the Armijo and Wolfe conditions. The Armijo condition is

\[h(t) ≤ h₀ + τ₀ t h'(0)\]

and the Wolfe condition is

\[h'(t) ≤ τ₁ h'(0).\]

Initially the step is increased trying to satisfy the Wolfe condition. Afterwards, only backtracking is performed in order to try to satisfy the Armijo condition. The final steplength may only satisfy Armijo's condition.

The output is the following:

t: the step length;

good_grad: whether g is the gradient at x + t * d;

ht: the model value at t, i.e., f(x + t * d);

nbk: the number of times the steplength was decreased to satisfy the Armijo condition, i.e., number of backtracks;

nbW: the number of times the steplength was increased to satisfy the Wolfe condition.

The following keyword arguments can be provided:

t: starting steplength (default 1);

τ₀: slope factor in the Armijo condition (default max(1e-4, √ϵₘ));

τ₁: slope factor in the Wolfe condition. It should satisfy τ₁ > τ₀ (default 0.9999);

bk_max: maximum number of backtracks (default 10);

bW_max: maximum number of increases (default 5);

verbose: whether to print information (default false).

Stats

SolverTools.GenericExecutionStats — Type

GenericExecutionStats(status, nlp; ...)

A GenericExecutionStats is a struct for storing output information of solvers. It contains the following fields:

status: Indicates the output of the solver. Use show_statuses() for the full list;

solution: The final approximation returned by the solver (default: []);

objective: The objective value at solution (default: Inf);

dual_feas: The dual feasibility norm at solution (default: Inf);

primal_feas: The primal feasibility norm at solution (default: 0.0 if uncontrained, Inf otherwise);

multipliers: The Lagrange multiplers wrt to the constraints (default: []);

multipliers_L: The Lagrange multiplers wrt to the lower bounds on the variables (default: []);

multipliers_U: The Lagrange multiplers wrt to the upper bounds on the variables (default: []);

iter: The number of iterations computed by the solver (default: -1);

elapsed_time: The elapsed time computed by the solver (default: Inf);

counters::NLPModels.NLSCounters: The Internal structure storing the number of functions evaluations;

solver_specific::Dict{Symbol,Any}: A solver specific dictionary.

The counters variable is a copy of nlp's counters, and status is mandatory on construction. All other variables can be input as keyword arguments.

Notice that GenericExecutionStats does not compute anything, it simply stores.

SolverTools.show_statuses — Function

show_statuses()

Show the list of available statuses to use with GenericExecutionStats.

Trust-Region

SolverTools.TrustRegionException — Type

Exception type raised in case of error.

SolverTools.AbstractTrustRegion — Type

AbstractTrustRegion

An abstract trust region type so that specific trust regions define update rules differently. Child types must have at least the following fields:

acceptance_threshold :: AbstractFloat

initial_radius :: AbstractFloat

radius :: AbstractFloat

ratio :: AbstractFloat

and the following function:

update!(tr, step_norm)

SolverTools.aredpred — Function

ared, pred = aredpred(nlp, f, f_trial, Δm, x_trial, step, slope)

Compute the actual and predicted reductions ∆f and Δm, where Δf = f_trial - f is the actual reduction is an objective/merit/penalty function, Δm = m_trial - m is the reduction predicted by the model m of f. We assume that m is being minimized, and therefore that Δm < 0.

SolverTools.acceptable — Function

acceptable(tr)

Return true if a step is acceptable

LinearOperators.reset! — Function

reset!(tr)

Reset the trust-region radius to its initial value

SolverTools.get_property — Function

A basic getter for AbstractTrustRegion instances. Should be overhauled when it's possible to overload getfield() and setfield!(). See https://github.com/JuliaLang/julia/issues/1974

SolverTools.set_property! — Function

A basic setter for AbstractTrustRegion instances.

SolverTools.update! — Function

update!(tr, step_norm)

Update the trust-region radius based on the ratio of actual vs. predicted reduction and the step norm.

SolverTools.TrustRegion — Type

Basic trust region type.

SolverTools.TRONTrustRegion — Type

Trust region used by TRON